DBZ-6907 Extract an abstract transformation class from ExtractNewRecordState
