+++ debezium-core/src/main/java/io/debezium/transforms/ExtractNewRecordState.java
@@ -6,47 +6,27 @@
 package io.debezium.transforms;
 
 import static io.debezium.data.Envelope.FieldName.OPERATION;
-import static io.debezium.data.Envelope.FieldName.SOURCE;
-import static io.debezium.data.Envelope.FieldName.TIMESTAMP;
-import static io.debezium.data.Envelope.FieldName.TRANSACTION;
-import static io.debezium.pipeline.txmetadata.TransactionMonitor.DEBEZIUM_TRANSACTION_DATA_COLLECTION_ORDER_KEY;
-import static io.debezium.pipeline.txmetadata.TransactionMonitor.DEBEZIUM_TRANSACTION_ID_KEY;
-import static io.debezium.pipeline.txmetadata.TransactionMonitor.DEBEZIUM_TRANSACTION_TOTAL_ORDER_KEY;
+import static io.debezium.transforms.ExtractNewRecordStateConfigDefinition.CONFIG_FIELDS;
 import static org.apache.kafka.connect.transforms.util.Requirements.requireStruct;
 
 import java.util.ArrayList;
-import java.util.Arrays;
-import java.util.Collections;
-import java.util.HashMap;
-import java.util.LinkedHashMap;
 import java.util.List;
 import java.util.Map;
-import java.util.Objects;
 import java.util.Optional;
-import java.util.regex.Pattern;
-import java.util.stream.Collectors;
 
 import org.apache.kafka.common.config.ConfigDef;
 import org.apache.kafka.connect.connector.ConnectRecord;
 import org.apache.kafka.connect.data.Schema;
 import org.apache.kafka.connect.data.SchemaBuilder;
 import org.apache.kafka.connect.data.Struct;
-import org.apache.kafka.connect.errors.ConnectException;
-import org.apache.kafka.connect.header.ConnectHeaders;
 import org.apache.kafka.connect.header.Header;
 import org.apache.kafka.connect.header.Headers;
-import org.apache.kafka.connect.transforms.ExtractField;
-import org.apache.kafka.connect.transforms.InsertField;
 import org.apache.kafka.connect.transforms.ReplaceField;
-import org.apache.kafka.connect.transforms.Transformation;
 import org.apache.kafka.connect.transforms.util.SchemaUtil;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
-import io.debezium.config.Configuration;
 import io.debezium.config.Field;
-import io.debezium.data.Envelope.Operation;
-import io.debezium.transforms.ExtractNewRecordStateConfigDefinition.DeleteHandling;
 import io.debezium.util.BoundedConcurrentHashMap;
 import io.debezium.util.Strings;
 
@@ -69,15 +49,12 @@ import io.debezium.util.Strings;
  * @param <R> the subtype of {@link ConnectRecord} on which this transformation will operate
  * @author Jiri Pechanec
  */
-public class ExtractNewRecordState<R extends ConnectRecord<R>> implements Transformation<R> {
+public class ExtractNewRecordState<R extends ConnectRecord<R>> extends AbstractExtractNewRecordState<R> {
 
     private static final Logger LOGGER = LoggerFactory.getLogger(ExtractNewRecordState.class);
 
-    private static final String PURPOSE = "source field insertion";
     private static final String EXCLUDE = "exclude";
     private static final int SCHEMA_CACHE_SIZE = 64;
-    private static final Pattern FIELD_SEPARATOR = Pattern.compile("\\.");
-    private static final Pattern NEW_FIELD_SEPARATOR = Pattern.compile(":");
 
     private static final Field DROP_FIELDS_HEADER = Field.create("drop.fields.header.name")
             .withDisplayName("Specifies a header that contains a list of field names to be removed")
@@ -105,70 +82,26 @@ public class ExtractNewRecordState<R extends ConnectRecord<R>> implements Transf
                     + "`true`: dropped fields are removed if the schema indicates its optional leaving the schemas unchanged, "
                     + "`false`: dropped fields are removed from the key/value schemas, regardless of optionality.");
 
-    private boolean dropTombstones;
     private String dropFieldsHeaderName;
     private boolean dropFieldsFromKey;
     private boolean dropFieldsKeepSchemaCompatible;
-    private DeleteHandling handleDeletes;
-    private List<FieldReference> additionalHeaders;
-    private List<FieldReference> additionalFields;
-    private String routeByField;
-    private final ExtractField<R> afterDelegate = new ExtractField.Value<>();
-    private final ExtractField<R> beforeDelegate = new ExtractField.Value<>();
-    private final InsertField<R> removedDelegate = new InsertField.Value<>();
-    private final InsertField<R> updatedDelegate = new InsertField.Value<>();
     private BoundedConcurrentHashMap<NewRecordValueMetadata, Schema> schemaUpdateCache;
-    private SmtManager<R> smtManager;
+
+    private final Field.Set configFields = CONFIG_FIELDS.with(
+            DROP_FIELDS_HEADER, DROP_FIELDS_FROM_KEY, DROP_FIELDS_KEEP_SCHEMA_COMPATIBLE);
 
     @Override
     public void configure(final Map<String, ?> configs) {
-        final Configuration config = Configuration.from(configs);
-        smtManager = new SmtManager<>(config);
-
-        final Field.Set configFields = Field.setOf(ExtractNewRecordStateConfigDefinition.DROP_TOMBSTONES,
-                ExtractNewRecordStateConfigDefinition.HANDLE_DELETES);
-        if (!config.validateAndRecord(configFields, LOGGER::error)) {
-            throw new ConnectException("Unable to validate config.");
-        }
-
-        dropTombstones = config.getBoolean(ExtractNewRecordStateConfigDefinition.DROP_TOMBSTONES);
-        handleDeletes = DeleteHandling.parse(config.getString(ExtractNewRecordStateConfigDefinition.HANDLE_DELETES));
-
-        String addFieldsPrefix = config.getString(ExtractNewRecordStateConfigDefinition.ADD_FIELDS_PREFIX);
-        String addHeadersPrefix = config.getString(ExtractNewRecordStateConfigDefinition.ADD_HEADERS_PREFIX);
-        additionalFields = FieldReference.fromConfiguration(addFieldsPrefix, config.getString(ExtractNewRecordStateConfigDefinition.ADD_FIELDS));
-        additionalHeaders = FieldReference.fromConfiguration(addHeadersPrefix, config.getString(ExtractNewRecordStateConfigDefinition.ADD_HEADERS));
-
-        String routeFieldConfig = config.getString(ExtractNewRecordStateConfigDefinition.ROUTE_BY_FIELD);
-        routeByField = routeFieldConfig.isEmpty() ? null : routeFieldConfig;
+        super.configure(configs);
 
         dropFieldsHeaderName = config.getString(DROP_FIELDS_HEADER);
         dropFieldsFromKey = config.getBoolean(DROP_FIELDS_FROM_KEY);
         dropFieldsKeepSchemaCompatible = config.getBoolean(DROP_FIELDS_KEEP_SCHEMA_COMPATIBLE);
-
-        Map<String, String> delegateConfig = new LinkedHashMap<>();
-        delegateConfig.put("field", "before");
-        beforeDelegate.configure(delegateConfig);
-
-        delegateConfig = new HashMap<>();
-        delegateConfig.put("field", "after");
-        afterDelegate.configure(delegateConfig);
-
-        delegateConfig = new HashMap<>();
-        delegateConfig.put("static.field", ExtractNewRecordStateConfigDefinition.DELETED_FIELD);
-        delegateConfig.put("static.value", "true");
-        removedDelegate.configure(delegateConfig);
-
-        delegateConfig = new HashMap<>();
-        delegateConfig.put("static.field", ExtractNewRecordStateConfigDefinition.DELETED_FIELD);
-        delegateConfig.put("static.value", "false");
-        updatedDelegate.configure(delegateConfig);
-
         schemaUpdateCache = new BoundedConcurrentHashMap<>(SCHEMA_CACHE_SIZE);
     }
 
     @Override
-    public R apply(final R record) {
+    public R doApply(final R record) {
         if (record.value() == null) {
             if (dropTombstones) {
                 LOGGER.trace("Tombstone {} arrived and requested to be dropped", record.key());
@@ -246,38 +179,16 @@ public class ExtractNewRecordState<R extends ConnectRecord<R>> implements Transf
         }
     }
 
-    private R setTopic(String updatedTopicValue, R record) {
-        String topicName = updatedTopicValue == null ? record.topic() : updatedTopicValue;
-
-        return record.newRecord(
-                topicName,
-                record.kafkaPartition(),
-                record.keySchema(),
-                record.key(),
-                record.valueSchema(),
-                record.value(),
-                record.timestamp());
+    @Override
+    public Iterable<Field> validateConfigFields() {
+        return configFields;
     }
 
-    /**
-     * Create an Headers object which contains the headers to be added.
-     */
-    private Headers makeHeaders(List<FieldReference> additionalHeaders, Struct originalRecordValue) {
-        Headers headers = new ConnectHeaders();
-
-        for (FieldReference fieldReference : additionalHeaders) {
-            // add "d" operation header to tombstone events
-            if (originalRecordValue == null) {
-                if (OPERATION.equals(fieldReference.field)) {
-                    headers.addString(fieldReference.getNewField(), Operation.DELETE.code());
-                }
-                continue;
-            }
-            Optional<Schema> schema = fieldReference.getSchema(originalRecordValue.schema());
-            schema.ifPresent(value -> headers.add(fieldReference.getNewField(), fieldReference.getValue(originalRecordValue), value));
-        }
-
-        return headers;
+    @Override
+    public ConfigDef config() {
+        final ConfigDef config = new ConfigDef();
+        Field.group(config, null, configFields.asArray());
+        return config;
     }
 
     private R addFields(List<FieldReference> additionalFields, R originalRecord, R unwrappedRecord) {
@@ -380,15 +291,6 @@ public class ExtractNewRecordState<R extends ConnectRecord<R>> implements Transf
         return fieldsToDrop;
     }
 
-    private Header getHeaderByName(R record, String headerName) {
-        for (Header header : record.headers()) {
-            if (header.key().equals(headerName)) {
-                return header;
-            }
-        }
-        return null;
-    }
-
     private Schema makeUpdatedSchema(List<FieldReference> additionalFields, Schema schema, Struct originalRecordValue) {
         // Get fields from original schema
         SchemaBuilder builder = SchemaUtil.copySchemaBasics(schema, SchemaBuilder.struct());
@@ -413,172 +315,5 @@ public class ExtractNewRecordState<R extends ConnectRecord<R>> implements Transf
 
     private Struct updateValue(FieldReference fieldReference, Struct updatedValue, Struct struct) {
         return updatedValue.put(fieldReference.getNewField(), fieldReference.getValue(struct));
-    }
-
-    @Override
-    public ConfigDef config() {
-        final ConfigDef config = new ConfigDef();
-        Field.group(config, null, ExtractNewRecordStateConfigDefinition.DROP_TOMBSTONES,
-                ExtractNewRecordStateConfigDefinition.HANDLE_DELETES, ExtractNewRecordStateConfigDefinition.ADD_FIELDS,
-                ExtractNewRecordStateConfigDefinition.ADD_HEADERS,
-                ExtractNewRecordStateConfigDefinition.ROUTE_BY_FIELD);
-        return config;
-    }
-
-    @Override
-    public void close() {
-        beforeDelegate.close();
-        afterDelegate.close();
-        removedDelegate.close();
-        updatedDelegate.close();
-    }
-
-    /**
-     * Represents a field that should be added to the outgoing record as a header
-     * attribute or struct field.
-     */
-    private static class FieldReference {
-
-        /**
-         * The struct ("source", "transaction") hosting the given field, or {@code null} for "op" and "ts_ms".
-         */
-        private final String struct;
-
-        /**
-         * The simple field name.
-         */
-        private final String field;
-
-        /**
-         * The name for the outgoing attribute/field, e.g. "__op" or "__source_ts_ms" when the prefix is "__"
-         */
-        private final String newField;
-
-        private FieldReference(String prefix, String field) {
-            String[] parts = NEW_FIELD_SEPARATOR.split(field);
-            String[] splits = FIELD_SEPARATOR.split(parts[0]);
-            this.field = splits.length == 1 ? splits[0] : splits[1];
-            this.struct = (splits.length == 1) ? determineStruct(this.field) : splits[0];
-
-            if (parts.length == 1) {
-                this.newField = prefix + (splits.length == 1 ? this.field : this.struct + "_" + this.field);
-            }
-            else if (parts.length == 2) {
-                this.newField = prefix + parts[1];
-            }
-            else {
-                throw new IllegalArgumentException("Unexpected field name: " + field);
-            }
-        }
-
-        /**
-         * Determines the struct hosting the given unqualified field.
-         */
-        private static String determineStruct(String simpleFieldName) {
-
-            switch (simpleFieldName) {
-                case DEBEZIUM_TRANSACTION_ID_KEY:
-                case DEBEZIUM_TRANSACTION_DATA_COLLECTION_ORDER_KEY:
-                case DEBEZIUM_TRANSACTION_TOTAL_ORDER_KEY:
-                    return TRANSACTION;
-                case OPERATION:
-                case TIMESTAMP:
-                    return null;
-                default:
-                    return SOURCE;
-
-            }
-        }
-
-        static List<FieldReference> fromConfiguration(String fieldPrefix, String addHeadersConfig) {
-            if (Strings.isNullOrEmpty(addHeadersConfig)) {
-                return Collections.emptyList();
-            }
-            else {
-                return Arrays.stream(addHeadersConfig.split(","))
-                        .map(String::trim)
-                        .map(field -> new FieldReference(fieldPrefix, field))
-                        .collect(Collectors.toList());
-            }
-        }
-
-        public String getNewField() {
-            return this.newField;
-        }
-
-        Object getValue(Struct originalRecordValue) {
-            Struct parentStruct = struct != null ? (Struct) originalRecordValue.getWithoutDefault(struct) : originalRecordValue;
-
-            // transaction is optional; e.g. not present during snapshotting atm.
-            return parentStruct != null ? getWithoutDefault(parentStruct, originalRecordValue) : null;
-        }
-
-        private Object getWithoutDefault(Struct parentStruct, Struct originalRecordValue) {
-            // In case field was added by other SMT and is in the main payload
-            // object.
-            return isInSchema(parentStruct.schema()) ? parentStruct.getWithoutDefault(field)
-                    : originalRecordValue.getWithoutDefault(field);
-        }
-
-        Optional<Schema> getSchema(Schema originalRecordSchema) {
-
-            Optional<org.apache.kafka.connect.data.Field> extractedField = getField(originalRecordSchema);
-            return extractedField.map(value -> SchemaUtil.copySchemaBasics(value.schema()).optional().build());
-
-        }
-
-        private Optional<org.apache.kafka.connect.data.Field> getField(Schema originalRecordSchema) {
-
-            Schema parentSchema = struct != null ? originalRecordSchema.field(struct).schema() : originalRecordSchema;
-
-            org.apache.kafka.connect.data.Field schemaField = parentSchema.field(field);
-
-            if (schemaField == null) {
-                LOGGER.debug("Field {} not found in {}. Trying in main payload", field, struct);
-                if (!isInSchema(originalRecordSchema)) {
-                    return Optional.empty();
-                }
-                schemaField = originalRecordSchema.field(field);
-            }
-
-            return Optional.of(schemaField);
-        }
-
-        private boolean isInSchema(Schema originalRecordSchema) {
-            return originalRecordSchema.field(field) != null;
-        }
-    }
-
-    private static class NewRecordValueMetadata {
-        private final Schema schema;
-        private final String operation;
-
-        NewRecordValueMetadata(Schema schema, String operation) {
-            this.schema = schema;
-            this.operation = operation;
-        }
-
-        @Override
-        public boolean equals(Object o) {
-            if (this == o) {
-                return true;
-            }
-            if (o == null || getClass() != o.getClass()) {
-                return false;
-            }
-            NewRecordValueMetadata metadata = (NewRecordValueMetadata) o;
-            return Objects.equals(schema, metadata.schema) &&
-                    Objects.equals(operation, metadata.operation);
-        }
-
-        @Override
-        public int hashCode() {
-            return Objects.hash(schema, operation);
-        }
-
-        @Override
-        public String toString() {
-            return "NewRecordValueMetadata{" + schema + ":" + operation + "}";
-        }
     }
 }
